Prevent a race condition that was triggered by subscribing to push notifications
Because CombineLatest was being used, we were gettings updates both when the toggle was enabled, and when the subscription was enabled when the permission request succeeded. This would result in two calls to the server to save the subscription and one would fail because the key already existed. Now we only save the subscription when the SwPush.subscription updates and not when requesting permissions succeeds.